finshed the backend. added middleware for route protection. fully functional CRUD with JWT athentification. need to clean up

+++ Travery_Media/mern-tutorial/backend/controllers/goalControllers.js
@@ -1,11 +1,13 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModels');
+const User = require('../models/userModels');
+
 // @desc Get Goals
 // @route GET /api/goals
 // @access Private
 
 const getGoals = asyncHandler(async (req, res) => {
-  const allGoals = await Goal.find();
+  const allGoals = await Goal.find({ user: req.user.id });
   res.status(200).json(allGoals);
 });
 
@@ -18,7 +20,10 @@ const setGoals = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error('Please add a text field');
   }
-  const createdGoal = await Goal.create({ text: req.body.text });
+  const createdGoal = await Goal.create({
+    text: req.body.text,
+    user: req.user.id,
+  });
 
   res.status(200).json(createdGoal);
 });
@@ -32,15 +37,36 @@ const updateGoals = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error('Please add a text field');
   }
-    const goal = await Goal.findById(req.params.id);
+  const goal = await Goal.findById(req.params.id);
 
-  if(!goal){
+  if (!goal) {
     res.status(400);
-    throw new Error('Goal not found')
+    throw new Error('Goal not found');
   }
-  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body,{new:true});
 
-  res.status(200).json({ message: `The goal with ${req.params.id} id has been updated`,updatedGoal});
+  const user = await User.findById(req.user.id);
+
+  //Check for user
+
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+  //Check to ensure logged in user matches the goal setting user
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
+  }
+
+  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+  });
+
+  res.status(200).json({
+    message: `The goal with ${req.params.id} id has been updated`,
+    updatedGoal,
+  });
 });
 
 // @desc Remove Goals
@@ -48,17 +74,32 @@ const updateGoals = asyncHandler(async (req, res) => {
 // @access Private
 
 const removeGoals = asyncHandler(async (req, res) => {
-
   const goal = await Goal.findById(req.params.id);
 
-  if(!goal){
+  if (!goal) {
     res.status(400);
-    throw new Error('Goal not found')
+    throw new Error('Goal not found');
   }
 
-  // const deletedGoal = await Goal.deleteOne({_id:req.params.id})
+  const user = await User.findById(req.user.id);
 
-  await goal.deleteOne({_id:req.params.id})
+  //Check for user
+
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+  //Check to ensure logged in user matches the goal setting user
+
+  
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
+  }
+  
+
+  await goal.deleteOne({ _id: req.params.id });
 
   res
     .status(200)
